makefile (GNU): compile first programs that do not require opencbm.
Otherwise, if opencbm is not available - it is not possible
to build the program that do not need openpbm.

Note that this modification is a quick shortcut, not resolving
the problem fully (ie. install rules will not work with not
all programs built...).
In general, improving configuration options and build
dependencies should allow to build things depending what
is needed. In particular, the tools that do not require
the openpbm driver, should not require it for building.